Array name declaration and CDATA support
Added a new way to call the deserialize method involving an
"options" object. The function can be called as usual with
an xml string and a callback. Added the possibility to call
with an options object containing two fields named "xml" and
"arrayNames". The first one shall contain the xmlstring, the
second one shall contain and array of strings corresponding
to object field names. If one of these names is encountered,
then the data element of the pair will be immediately
initialized as a string instead of an object. Added CDATA
support as well.